Refactored PowBlock into CheckpointBlock and TransactionBlock
